+++ src/array/codec/array_to_array/transpose/transpose_configuration.rs
@@ -1,5 +1,7 @@
 use derive_more::{Display, From};
-use serde::{ser::SerializeSeq, Deserialize, Serialize};
+use serde::{Deserialize, Serialize};
+
+use super::{validate_permutation, InvalidPermutationError};
 
 /// A wrapper to handle various versions of Transpose codec configuration parameters.
 #[derive(Serialize, Deserialize, Clone, Eq, PartialEq, Debug, Display, From)]
@@ -14,7 +16,7 @@ pub enum TransposeCodecConfiguration {
 #[serde(deny_unknown_fields)]
 #[display(fmt = "{}", "serde_json::to_string(self).unwrap_or_default()")]
 pub struct TransposeCodecConfigurationV1 {
-    /// An array of integers specifying the permutation, "C", or "F".
+    /// The transpose order defining how to permute the array.
     pub order: TransposeOrder,
 }
 
@@ -27,59 +29,35 @@ impl TransposeCodecConfigurationV1 {
 }
 
 /// The transpose order defining how to permute the array.
-#[derive(Clone, Eq, PartialEq, Debug)]
-pub enum TransposeOrder {
-    /// The string `"C"`, equivalent to specifying the identity permutation `0`, `1`, …, `n-1`. This makes the codec a no-op.
-    C,
+///
+/// An array of integers specifying a permutation of 0, 1, …, n-1, where n is the number of dimensions in the decoded chunk representation provided as input to this codec.
+#[derive(Serialize, Clone, Eq, PartialEq, Debug)]
+pub struct TransposeOrder(pub Vec<usize>);
 
-    /// The string `"F"`, equivalent to specifying the permutation `n-1`, …, `1`, `0`.
-    F,
-
-    /// An array of integers specifying a permutation of 0, 1, …, n-1, where n is the number of dimensions in the decoded chunk representation provided as input to this codec.
-    Permutation(Vec<usize>),
-}
-
-impl serde::Serialize for TransposeOrder {
-    fn serialize<S: serde::Serializer>(&self, s: S) -> Result<S::Ok, S::Error> {
-        match &self {
-            Self::C => s.serialize_str("C"),
-            Self::F => s.serialize_str("F"),
-            Self::Permutation(permutation) => {
-                let mut seq = s.serialize_seq(Some(permutation.len()))?;
-                for v in permutation {
-                    seq.serialize_element(&v)?;
-                }
-                seq.end()
-            }
+impl TransposeOrder {
+    /// Create a new [`TransposeOrder`].
+    ///
+    /// # Errors
+    /// Returns [`InvalidPermutationError`] if the permutation order is invalid.
+    pub fn new(order: &[usize]) -> Result<Self, InvalidPermutationError> {
+        if validate_permutation(order) {
+            Ok(Self(order.to_vec()))
+        } else {
+            Err(InvalidPermutationError::from(order.to_vec()))
         }
     }
 }
 
 impl<'de> serde::Deserialize<'de> for TransposeOrder {
     fn deserialize<D: serde::Deserializer<'de>>(d: D) -> Result<Self, D::Error> {
-        let value = serde_json::Value::deserialize(d)?;
-        match value {
-            serde_json::Value::String(string) => {
-                if string == "C" {
-                    return Ok(Self::C);
-                } else if string == "F" {
-                    return Ok(Self::F);
-                }
-            }
-            serde_json::Value::Array(array) => {
-                if array.iter().all(serde_json::Value::is_u64) {
-                    let permutation: Vec<usize> = array
-                        .iter()
-                        .map(|v| v.as_u64().unwrap().try_into().unwrap())
-                        .collect();
-                    return Ok(Self::Permutation(permutation));
-                }
-            }
-            _ => {}
+        let permutation = Vec::<usize>::deserialize(d)?;
+        if validate_permutation(&permutation) {
+            Ok(Self(permutation))
+        } else {
+            Err(serde::de::Error::custom(
+                "transpose order must be an array of integers specifying a permutation of 0, 1, …, n-1, where n is the number of dimensions",
+            ))
         }
-        Err(serde::de::Error::custom(
-            "transpose order must be C, F, or an array of integers specifying a permutation",
-        ))
     }
 }
 
@@ -87,37 +65,26 @@ impl<'de> serde::Deserialize<'de> for TransposeOrder {
 mod tests {
     use super::*;
 
-    const JSON_C: &str = r#"{
-        "order": "C"
-    }"#;
-
-    const JSON_F: &str = r#"{
-        "order": "F"
-    }"#;
-
-    const JSON_ARRAY: &str = r#"{
-        "order": [0, 2, 1]
-    }"#;
-
-    #[test]
-    fn codec_transpose_c() {
-        serde_json::from_str::<TransposeCodecConfiguration>(JSON_C).unwrap();
-    }
-
-    #[test]
-    fn codec_transpose_f() {
-        serde_json::from_str::<TransposeCodecConfiguration>(JSON_F).unwrap();
-    }
-
     #[test]
     fn codec_transpose_array() {
-        serde_json::from_str::<TransposeCodecConfiguration>(JSON_ARRAY).unwrap();
+        let json = r#"{
+            "order": [0, 2, 1]
+        }"#;
+        serde_json::from_str::<TransposeCodecConfiguration>(json).unwrap();
     }
 
     #[test]
-    fn codec_transpose_invalid() {
+    fn codec_transpose_invalid1() {
         let json = r#"{
             "order": ""
+        }"#;
+        assert!(serde_json::from_str::<TransposeCodecConfiguration>(json).is_err());
+    }
+
+    #[test]
+    fn codec_transpose_invalid2() {
+        let json = r#"{
+            "order": [0, 2]
         }"#;
         assert!(serde_json::from_str::<TransposeCodecConfiguration>(json).is_err());
     }